@@ -1,15 +1,15 @@
 apply plugin: 'com.android.application'
 
 android {
-    compileSdkVersion 28
-    buildToolsVersion "28.0.3"
+    compileSdkVersion 29
+    buildToolsVersion "29.0.3"
 
     defaultConfig {
         applicationId "com.thomaskuenneth.android.birthday"
         minSdkVersion 14
-        targetSdkVersion 28
-        versionCode = 20103
-        version = "2.1.2"
+        targetSdkVersion 29
+        versionCode = 20104
+        version = "2.1.3"
     }
 
     buildTypes {
@@ -20,6 +20,6 @@ android {
     }
 
     dependencies {
-        implementation 'com.android.support:appcompat-v7:28.0.0'
+        implementation 'androidx.appcompat:appcompat:1.1.0'
     }
 }
